+++ docs/release-notes.md
@@ -3,6 +3,7 @@ Jupiter release notes
 ### 2018-06-06:
 
 - DefaultInvokeFuture#roundFutures初始容量配置, 避免频繁扩容
+- 优化[DefaultInvokeFuture#TimeoutScanner优化](https://github.com/fengjiachun/Jupiter/issues/62)
 
 ### 2018-04-13: version 1.2.21